# Speed up message extraction by using one rule (one pass) instead of 4