--- a/lib/defaultConfig.js
+++ b/lib/defaultConfig.js
@@ -1,9 +1,6 @@
 module.exports = {
     stringsRules: [
-            /intl\.formatMessage\s*?\(\s*?(['`"])([\s\S]+?)\1\s*?(?=[,\)])/g,
-        /intl\.formatHTMLMessage\s*?\(\s*?(['`"])([\s\S]+?)\1\s*?(?=[,\)])/g,
-              /\{\{formatMessage\s*?\(\s*?(['`"])([\s\S]+?)\1\s*?(?=[,\)])/g,
-          /\{\{formatHTMLMessage\s*?\(\s*?(['`"])([\s\S]+?)\1\s*?(?=[,\)])/g
+        /(?:intl\.|\{\{)format(?:HTML)?Message\s*?\(\s*?(['`"])([\s\S]+?)\1\s*?(?=[,\)])/g
     ],
     scopeRules: [
         /\/\/\s*intlScope:\s*([\-\w\.]+)/g,