Running `tau-extract-gettext.cmd scripts/tau`
Time with 4 rules: getScope: 33s, tokenize: 86s, tokenizeJsx: 321ms
Time with 1 rule: getScope: 33s, tokenize: 20s, tokenizeJsx: 313ms